requirements.txt
brotlipy==0.7.0
certifi==2020.4.5.1
cffi==1.14.0
click==7.1.2
dash==1.12.0
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
pandas==1.0.3
plotly==4.7.1
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
retrying==1.3.3
six==1.14.0
Werkzeug==1.0.1
wincertstore==0.2
mysql-connector-python==8.0.20